requirements.txt
# Core Data Science (Critical: numpy must be installed first)
numpy==1.23.5  # Specific version for compatibility
pandas>=1.3.0
scikit-learn>=0.24.2

# Time Series Models (pmdarima must be installed after numpy)
statsmodels>=0.13.0
prophet>=1.1.0   # For Prophet models
torch>=1.9.0     # For LSTM models
pmdarima>=2.0.0  # For ARIMA models - install after numpy

# MLOps Tools
mlflow>=2.0.0
prefect>=2.0.0
streamlit>=1.0.0
optuna>=3.0.0    # For hyperparameter optimization

# Monitoring & Visualization
plotly>=5.0.0
matplotlib>=3.4.0
seaborn>=0.11.0
psutil>=5.8.0

# API & Dashboard
fastapi>=0.68.0
uvicorn>=0.15.0
streamlit-autorefresh>=0.0.1
streamlit-metrics>=0.1.0

# Testing & Development
pytest>=6.2.5
pytest-cov>=2.12.0
black>=21.7b0
flake8>=3.9.0
isort>=5.9.0

# Utilities
python-dotenv>=0.19.0
pyyaml>=5.4.0
joblib>=1.0.0
tqdm>=4.62.0
watchdog>=2.1.0

# Optional: Deep Learning GPU Support
# torch-cuda  # Uncomment if GPU support is needed 